+++ packages/calcite-components/src/components/dialog/dialog.tsx
@@ -620,9 +620,16 @@ export class Dialog
       dragPosition: { x, y },
       resizePosition,
       transitionEl,
+      dragEnabled,
+      resizable,
     } = this;
 
     if (!transitionEl) {
+      return;
+    }
+
+    if (!dragEnabled && !resizable) {
+      transitionEl.style.transform = null;
       return;
     }
 
@@ -631,7 +638,8 @@ export class Dialog
     const translateX = Math.round(x + left + right);
     const translateY = Math.round(y + top + bottom);
 
-    transitionEl.style.transform = `translate(${translateX}px, ${translateY}px)`;
+    transitionEl.style.transform =
+      translateX || translateY ? `translate(${translateX}px, ${translateY}px)` : null;
   }
 
   private updateSize({