+++ packages/calcite-components-react/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.14.0-next.7](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@2.14.0-next.6...@esri/calcite-components-react@2.14.0-next.7) (2024-10-08)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [2.14.0-next.6](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@2.14.0-next.5...@esri/calcite-components-react@2.14.0-next.6) (2024-10-07)
 
 **Note:** Version bump only for package @esri/calcite-components-react